Can now calculate tta and brix amount where there is too much acid

+++ src/main/java/domain/KombuchaTank.java
@@ -1,34 +1,32 @@
 package domain;
 
-import java.math.BigDecimal;
-
 public class KombuchaTank implements Tank {
 
-    private BigDecimal volume;
-    private BigDecimal brixValue;
-    private BigDecimal ttaValue;
+    private double volume;
+    private double brixValue;
+    private double ttaValue;
 
-    public KombuchaTank(BigDecimal volume, BigDecimal brixValue, BigDecimal ttaValue) {
+    public KombuchaTank(double volume, double brixValue, double ttaValue) {
         this.volume = volume;
         this.brixValue = brixValue;
         this.ttaValue = ttaValue;
     }
 
     @Override
-    public BigDecimal getVolume() {
+    public double getVolume() {
         return volume;
     }
 
     @Override
-    public BigDecimal getValueOf(PROPERTY_TYPE propertyType) {
+    public double getValueOf(PROPERTY_TYPE propertyType) {
         return PROPERTY_TYPE.BRIX.equals(propertyType) ? getBrixValue() : getTtaValue();
     }
 
-    private BigDecimal getBrixValue() {
+    private double getBrixValue() {
         return brixValue;
     }
 
-    private BigDecimal getTtaValue() {
+    private double getTtaValue() {
         return ttaValue;
     }
